Make clang build linkable: Part 2 (#100)
* Stub sc_sub.c

* Stub SDK functions

* Stub some funcs

* Stub bg.c

* Stub more files

* Stub more files

+++ include/sf33rd/Source/Game/workuser.h
@@ -4,8 +4,8 @@
 #include "types.h"
 
 extern s16 Opening_Now;            // size: 0x2, address: 0x579DB0
-extern u8 plsw_01[2];              // size: 0x4, address: 0x579DD0
-extern u8 plsw_00[2];              // size: 0x4, address: 0x579DD4
+extern u16 plsw_01[2];              // size: 0x4, address: 0x579DD0
+extern u16 plsw_00[2];              // size: 0x4, address: 0x579DD4
 extern u16 VS_Win_Record[2];       // size: 0x4, address: 0x579DD8
 extern u16 IO_Result;              // size: 0x2, address: 0x579DDC
 extern u16 Keep_Grade[2];          // size: 0x4, address: 0x579DE0
